// src/app/models/objectInfo.ts
export interface ObjectInfo {
    assortmentModule: string;
    gainings12Mrsd: number;
    wdpercentSerbia: number;
    wdpercentSector: number;
    wdpercentCustomer: number;
    gainingsVs12mpercent: number;
    registersNumber: number;
    shelfSpaceM: number;
    companyShelfSpaceM: number;
    companyShelfSpacePercent: number;
  }
  